# openapi.yaml
openapi: 3.0.2
info:
  description: |
    Customers focuses on solving authentic identification of humans who are legally able to hold and transfer currency within the US. Primarily this project solves [Know Your Customer](https://en.wikipedia.org/wiki/Know_your_customer) (KYC), [Customer Identification Program](https://en.wikipedia.org/wiki/Customer_Identification_Program) (CIP), [Office of Foreign Asset Control](https://www.treasury.gov/about/organizational-structure/offices/Pages/Office-of-Foreign-Assets-Control.aspx) (OFAC) checks and verification workflows to comply with US federal law and ensure authentic transfers. Customers has an objective to be a service for detailed due diligence on individuals and companies for Financial Institutions and services in a modernized and extensible way.

    Customer phone numbers and addresses are stored and partially used in KYC/OFAC validation. Arbitrary key/value pairs can be stored for a Customer. Documents and Disclaimers, and their acknowledgement are also stored under a Customer as they're accepted. Bank Accounts, which can be validated with micro-deposits currently, are stored under each Customer.

    ![](https://raw.githubusercontent.com/adamdecaf/customers/create-accounts/docs/images/customer.png)
  version: v1
  title: Customers API
  contact:
    url: https://github.com/moov-io/customers
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8087
...
paths:
...
  /customers:
    post:
      tags: [Customers]
      summary: Create customer
      description: Create a Customer object from the given details of a human or business
      operationId: createCustomer
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomer'
      responses:
        '200':
          description: Customer was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Customer was not created, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /customers/{customerID}:
    get:
      tags: [Customers]
      summary: Retrieve customer
      description: Get the Customer object and metadata for the customerID.
      operationId: getCustomer
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID that identifies this Customer
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      responses:
        '200':
          description: A customer objects for the supplied customerID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: No Customer with the specified customerID was found
  /customers/{customerID}/address:
    post:
      tags: [Customers]
      summary: Add customer address
      description: Add an Address onto an existing Customer record
      operationId: addCustomerAddress
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to add the address onto
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerAddress'
      responses:
        '200':
          description: A customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Customer address was not updated, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /customers/{customerID}/metadata:
    put:
      tags: [Customers]
      summary: Update customer metadata
      description: Replace the metadata object for a customer. Metadata is a map of unique keys associated to values to act as foreign key relationships or arbitrary data associated to a Customer.
      operationId: replaceCustomerMetadata
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to add the metadata onto
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerMetadata'
      responses:
        '200':
          description: A customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Customer metadata was not updated, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /customers/{customerID}/status:
    put:
      tags: [Customers]
      summary: Update customer status
      description: Update the status for a customer, which can only be updated by authenticated users with permissions.
      operationId: updateCustomerStatus
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to update the CustomerStatus
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerStatus'
      responses:
        '200':
          description: A customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Customer status was not updated, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
...
components:
  schemas:
    CreateCustomer:
      type: object
      properties:
        firstName:
          type: string
          maximum: 64
          description: Given Name or First Name
          example: Robert
        middleName:
          type: string
          maximum: 64
          description: Middle Name
          example: Flex
        lastName:
          type: string
          description: Surname or Last Name
          example: Smith
        nickName:
          type: string
          description: Name Customer is preferred to be called
          example: Bob
        suffix:
          type: string
          description: Customers name suffix. "Jr", "PH.D."
        type:
          $ref: '#/components/schemas/CustomerType'
        birthDate:
          type: string
          format: date-time
          description: Legal date of birth
          example: '2016-08-29T09:12:33.001Z'
        email:
          type: string
          maximum: 500
          description: Primary email address of customer name@domain.com
        SSN:
          type: string
          description: Customer Social Security Number (SSN)
          example: 111-11-1111
        phones:
          type: array
          items:
            $ref: '#/components/schemas/CreatePhone'
        addresses:
          type: array
          minimum: 1
          items:
            $ref: '#/components/schemas/CreateCustomerAddress'
        metadata:
          type: object
          description: Map of unique keys associated to values to act as foreign key relationships or arbitrary data associated to a Customer.
          additionalProperties:
            type: string
          example:
            paygateID: "23beb5fd"
      required:
        - firstName
        - lastName
        - birthDate
        - email
        - addresses
        # - SSN # TODO(adam): required, but we need to impl storage and crypto
        # - phones
        # - addresses
    Customer:
      type: object
      properties:
        customerID:
          type: string
          description: The unique identifier for the customer who owns the account
          example: e210a9d6
        firstName:
          type: string
          maximum: 64
          description: Given Name or First Name
          example: Robert
        middleName:
          type: string
          maximum: 64
          description: Middle Name
          example: Flex
        lastName:
          type: string
          description: Surname or Last Name
          example: Smith
        nickName:
          type: string
          description: Name Customer is preferred to be called
          example: Bob
        suffix:
          type: string
          description: Customers name suffix. "Jr", "PH.D."
        type:
          $ref: '#/components/schemas/CustomerType'
        birthDate:
          type: string
          format: date-time
          description: Legal date of birth
          example: '2016-08-29T09:12:33.001Z'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        email:
          type: string
          maximum: 500
          description: Primary email address of customer name@domain.com
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/CustomerAddress'
        metadata:
          type: object
          description: Map of unique keys associated to values to act as foreign key relationships or arbitrary data associated to a Customer.
          additionalProperties:
            type: string
          example:
            paygateID: "23beb5fd"
        createdAt:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'
        lastModified:
          type: string
          format: date-time
          description: Last time the object was modified
          example: '2016-08-29T09:12:33.001Z'
    CustomerType:
      type: string
      description: Note if this Customer represents an individual or business
      enum:
        - Individual
        - Business
    CustomerStatus:
      type: string
      description: State of the customer
      enum:
        - Deceased
        - Rejected
        - Unknown
        - ReceiveOnly
        - Verified
    CustomerMetadata:
      properties:
        metadata:
          type: object
          description: Map of unique keys associated to values to act as foreign key relationships or arbitrary data associated to a Customer.
          additionalProperties:
            type: string
          example:
            paygateID: "23beb5fd"
      required:
        - metadata
    UpdateCustomerStatus:
      properties:
        comment:
          type: string
          description: Free form comment about the customer status update
          example: Customer was approved from KYC confirmation
        status:
          $ref: '#/components/schemas/CustomerStatus'
      required:
        - status
    CreatePhone:
      properties:
        number:
          type: string
          description: phone number
          example: "+1.818.555.1212"
        type:
          type: string
          enum:
            - Home
            - Mobile
            - Work
      required:
        - number
        - type
...
    CreateCustomerAddress:
      type: object
      properties:
        type:
          type: string
          enum:
            - Primary
            - Secondary
        address1:
          type: string
          description: First line of the address
        address2:
          type: string
          description: Second line of the address
        city:
          type: string
        state:
          type: string
          minimum: 2
          maximum: 2
          description: two charcer code of US state
        postalCode:
          type: string
        country:
          type: string
          enum:
            - US
      required:
        - type
        - address1
        - address2
        - city
        - state
        - postalCode
        - country
...